feat: updateFavoriteCoin 메소드 추가 및 insert, delete 메소드 제거, FavoriteCoinStorage find 메소드 추가 #64
- 현재 uuid가 있는지 확인하고 추가 삭제하도록 기능 변경

+++ Coin/Search/SearchCoinsViewModel.swift
@@ -37,11 +37,12 @@ final class SearchViewModel {
         return favoriteCoinRepository.fetch()
     }
     
-    func insertFavoriteCoin(from uuid: String) {
-        favoriteCoinRepository.insert(uuid: uuid)
-    }
-    
-    func deleteFavoriteCoin(from uuid: String) {
-        favoriteCoinRepository.delete(uuid: uuid)
+    func updateFavoriteCoin(from uuid: String) {
+        let findUUID = favoriteCoinRepository.find(uuid: uuid)
+        if findUUID {
+            favoriteCoinRepository.delete(uuid: uuid)
+        } else {
+            favoriteCoinRepository.insert(uuid: uuid)
+        }
     }
 }
